+++ source/runtime.js
@@ -4,7 +4,7 @@
 // Local
 const { status } = require('./log')
 const { isSpecified } = require('./string')
-const { spawn, write, unlink, exists, rename } = require('./fs')
+const { spawn, write, unlink, exists, rename, contains } = require('./fs')
 const { readPackage, writePackage } = require('./package')
 const { versionComparator } = require('./version')
 const { getAnswers } = require('./answers')
@@ -44,22 +44,45 @@ function addLatest (array) {
 
 // Update engines
 async function updateEngines (state) {
-	const { answers, supportedNodeVersions, nodeVersions } = state
+	const { answers, supportedNodeVersions, nodeVersions, packageData } = state
+	const nodeEditions = state.nodeEditions
+	let minimumPassingVersion = null
 
 	// =================================
 	// run each edition against the supported node version
 	// to fetch the engines for each edition
 
-	const versionsAlreadySupported = new Set()
-	const uselessEditions = []
+	if (nodeEditions.length === 0) {
+		// this can be the case if it is a website
+		status('determining engines for project...')
+		const versions = new Versions(nodeVersions)
+		await versions.load()
+		await versions.install()
+		const numbers = versions.map((version) => version.version)
+		await versions.test('npm test')
+		const passed = versions.json.passed || []
+		if (passed.length === 0) {
+			console.error(versions.messages.join('\n\n'))
+			throw new Error(`There were no node versions [${numbers.join(', ')}] which the project's tests passed`)
+		}
+		else {
+			minimumPassingVersion = passed[0]
+			packageData.engines.node = nodeMajorVersions(passed).join(' || ')
+		}
+		status(`...determined engines for project as [${state.engines.node}] against [${numbers.join(', ')}]`)
+	}
+	else {
+		const versionsAlreadySupported = new Set()
+		let recompile = false
 
-	/* eslint no-loop-func:0 */
-	for (const edition of state.editions) {
-		if (edition.engines && edition.engines.node) {
+		/* eslint no-loop-func:0 */
+		for (const edition of nodeEditions) {
 			status(`determining engines for edition [${edition.directory}]...`)
+
+			// run the test for the edition, and determine the results
 			const test = answers.docpadPlugin
 				? `docpad-plugintester --edition=${edition.directory}`
-				: `node --harmony ./${pathUtil.join(edition.directory || '.', edition.testEntry)} --joe-reporter=console`
+				: `node --harmony ./${pathUtil.join(edition.directory || '.', edition.test)} --joe-reporter=console`
 			const versions = new Versions(supportedNodeVersions.concat((edition.targets && edition.targets.node) || []))
 			await versions.load()
 			await versions.install()
@@ -73,124 +96,83 @@ async function updateEngines (state) {
 				edition.engines.node = true
 			}
 
-			// Update the sets
+			// update the sets
 			const passedUnique = passed.filter((version) => versionsAlreadySupported.has(nodeMajorVersion(version)) === false)
 			const failedUnique = failed.filter((version) => versionsAlreadySupported.has(nodeMajorVersion(version)) === false)
 			const failedRequired = edition.engines.node === true
 				? []
 				: failedUnique.filter((version) => semver.satisfies(version, edition.engines.node))
-			const supported = edition.engines.node === true
+			const targetedUnique = edition.engines.node === true
 				? passedUnique
 				: passedUnique.filter((version) => semver.satisfies(version, edition.engines.node))
 
-			console.log(`passed: ${passed.join(', ')}\nunique: ${passedUnique.join(', ')}\nsupported: ${supported.join(', ')}\nfailed: ${failed.join(', ')}\nunique: ${failedUnique.join(', ')}\nrequired: ${failedRequired.join(', ')}`)
+			// log the results
+			console.log([
+				`passed:    ${passed.join(', ')}`,
+				`unique:    ${passedUnique.join(', ')}`,
+				`failed:    ${failed.join(', ')}`,
+				`unique:    ${failedUnique.join(', ')}`,
+				`required:  ${failedRequired.join(', ')}`,
+				`supports:  ${targetedUnique.join(', ')}`
+			].join('\n'))
 
+			// error if unsuccessful
 			if (passed.length === 0) {
 				console.error(versions.messages.join('\n\n'))
 				throw new Error(`The edition [${edition.directory}] had no node versions [${numbers.join(', ')}] which its tests passed`)
 			}
 
+			// error if unsuccesful for required
 			if (failedRequired.length) {
 				console.error(versions.messages.join('\n\n'))
 				throw new Error(`The edition [${edition.directory}] with engines [${edition.engines.node}] needed to support [${failedRequired.join(', ')}] but failed on [${failed.join(', ')}]`)
 			}
 
-			if (edition.engines.node === true) {
-				edition.engines.node = '>=' + nodeMajorVersion(passed[0])
-			}
-			else {
-				edition.engines.node = nodeMajorVersions(passed).join(' || ')
-			}
+			// make engines the passed versions
+			edition.engines.node = nodeMajorVersions(passed).join(' || ')
 
-			if (supported.length === 0) {
-				console.log(`The edition [${edition.directory}] had no unique node versions that it supported, so will been trimmed`)
-				uselessEditions.push(edition)
+			// if edition is redundant, mark it as inactive, and mark runtime as needing to be recompiled
+			if (targetedUnique.length === 0) {
+				console.log(`The edition [${edition.directory}] had no unique node versions that it targeted, so will been trimmed`)
+				edition.active = false
+				recompile = true
 				continue
 			}
 
-			if (edition.directory === 'source' && state.useEditionAutoloader === false) {
-				// if we are not using the edition autoloader, then we don't care what versions the source edition supports
-			}
-			else {
-				passedUnique.forEach((version) => versionsAlreadySupported.add(nodeMajorVersion(version)))
-			}
+			// add the unique versions to the list
+			passedUnique.forEach((version) => versionsAlreadySupported.add(nodeMajorVersion(version)))
+
+			// log
 			status(`...determined engines for edition [${edition.directory}] as [${edition.engines.node}] against [${numbers.join(', ')}]`)
 		}
-	}
 
-	// =================================
-	// update the package.json file
-
-	await writePackage(state)
-
-	// =================================
-	// trim useless editions
-
-	if (state.useEditionAutoloader === false) {
-		// trim compiled editions that are supported by later edition
-		const nodeEditions = state.editions.filter((edition) => edition.engines.node)
-		for (let index = 0; index < nodeEditions.length - 1; index++) {
-			const edition = nodeEditions[index]
-			const nextEdition = nodeEditions[index + 1]
-			if (edition.directory !== 'source' && nextEdition.engines.node.includes(edition.engines.node) && uselessEditions.includes(edition) === false) {
-				console.log(`The edition [${edition.directory}] had no unique node versions that it supported, so will been trimmed`)
-				uselessEditions.push(edition)
-			}
+		// if there has been an editions change, try again with an updated runtime
+		if (recompile) {
+			return await updateRuntime(state)
 		}
-	}
 
-	if (uselessEditions.length) {
-		// need to do this, as we have to trim the editions and the autoloader
-		// no need for this to update the scripts and babel, as they are done in writePackage, not updateRuntime
-
-		status(`removing useless editions ${uselessEditions.map((edition) => edition.directory).join(', ')}...`)
-
-		state.editions = state.editions.filter((edition) => uselessEditions.includes(edition) === false)
-
-		/* await Promise.all(
-			uselessEditions.map((edition) => unlink(edition.directory))
-		) - unlink is not suffecient for nested dirs */
-
-		status('...removed useless editions')
-
-		return await updateRuntime(state)
+		// get the first passing version
+		minimumPassingVersion = Array.from(versionsAlreadySupported.values()).sort(versionComparator)[0]
 	}
 
 	// =================================
-	// figure out engines.node
+	// update engines.node
 
-	// @todo skip if state.editions.length === 1
-	if (state.editions.length >= 1) {
-		status('determining engines for project...')
-		const versions = new Versions(nodeVersions)
-		await versions.load()
-		await versions.install()
-		const numbers = versions.map((version) => version.version)
-		await versions.test('npm test')
-		const passed = versions.json.passed || []
-		if (passed.length === 0) {
-			console.error(versions.messages.join('\n\n'))
-			throw new Error(`There were no node versions [${numbers.join(', ')}] which the project's tests passed`)
-		}
-		status('...determined engines for project')
-
-		// check if minimum supported node version is still the minimum supported node version
-		if (versionComparator(passed[0], answers.minimumSupportNodeVersion) !== 0) {
-			console.error(versions.messages.join('\n\n'))
-			const message = [
-				`The project actually supports the minimum node version ${passed[0]} which is different than your specified minimum supported node version ${answers.minimumSupportNodeVersion}`,
-				'What would you like to do?'
-			].join('\n')
-			const query = await getAnswers([{
-				name: 'action',
-				type: 'list',
-				choices: ['ignore', 'fail'],
-				validate: isSpecified,
-				message
-			}])
-			if (query.action === 'fail') {
-				return process.exit(1)
-			}
+	// check if minimum supported node version is still the minimum supported node version
+	if (versionComparator(minimumPassingVersion, answers.minimumSupportNodeVersion) !== 0) {
+		const message = [
+			`The project actually supports the minimum node version ${minimumPassingVersion} which is different than your specified minimum supported node version ${answers.minimumSupportNodeVersion}`,
+			'What would you like to do?'
+		].join('\n')
+		const query = await getAnswers([{
+			name: 'action',
+			type: 'list',
+			choices: ['ignore', 'fail'],
+			validate: isSpecified,
+			message
+		}])
+		if (query.action === 'fail') {
+			return process.exit(1)
 		}
 	}
 
@@ -201,45 +183,31 @@ async function updateEngines (state) {
 }
 
 async function scaffoldEditions (state) {
-	const { editions, packageData, answers } = state
-	state.useEditions = editions.length
-	state.useEditionAutoloader = editions.length > 1 && answers.name !== 'editions'
-	if (state.useEditions) {
-		// source
-		const sourceEdition = editions[0]
-		const sourceMainEntry = sourceEdition.entry
-		const sourceTestEntry = sourceEdition.testEntry
-		const sourceMainPath = pathUtil.join(sourceEdition.directory || '.', sourceMainEntry)
-		const sourceTestPath = pathUtil.join(sourceEdition.directory || '.', sourceTestEntry)
-
-		// compiled (use slice to prevent reverse changing the original array)
-		const nodeEdition = editions.slice().reverse().find((edition) => edition.engines && edition.engines.node) || sourceEdition
-		const nodeMainEntry = nodeEdition.entry
-		const nodeTestEntry = nodeEdition.testEntry
-		const nodeMainPath = pathUtil.join(nodeEdition.directory || '.', nodeMainEntry)
-		const nodeTestPath = pathUtil.join(nodeEdition.directory || '.', nodeTestEntry)
+	const { activeEditions, packageData, answers } = state
+	if (activeEditions.length) {
+		// fetch
+		const sourceEdition = state.sourceEdition
+		const compiledEdition = state.compiledEdition || sourceEdition
 
 		// log
 		status('scaffolding edition files...')
 
-		// scaffold directories
+		// scaffold edition directories
 		await spawn(['mkdir', '-p'].concat(
-			editions.map(
+			activeEditions.map(
 				(edition) => edition.directory || '.'
 			)
 		))
 
 		// move or scaffold edition main path if needed
-		const sourceMainPathExists = await exists(sourceMainPath)
-		const sourceMainEntryExists = await exists(sourceMainEntry)
-		if (!sourceMainPathExists) {
+		if ((await exists(sourceEdition.mainPath)) === false) {
 			// edition entry doesn't exist, but the root entry does
-			if (sourceMainEntryExists) {
-				await rename(sourceMainEntry, sourceMainPath)
+			if (await exists(sourceEdition.main)) {
+				await rename(sourceEdition.main, sourceEdition.mainPath)
 			}
 			// edition entry doesn't exist, but it is a docpad plugin
 			else if (answers.docpadPlugin) {
-				write(sourceMainPath, [
+				write(sourceEdition.mainPath, [
 					"'use strict'",
 					'',
 					"module.exports = class MyPlugin extends require('docpad-baseplugin') {",
@@ -250,21 +218,19 @@ async function scaffoldEditions (state) {
 				].join('\n'))
 			}
 			// edition entry doesn't exist, so create an empty file
-			else await spawn(['touch', sourceMainPath])
+			else await spawn(['touch', sourceEdition.mainPath])
 		}
 
 		// move or scaffold edition test path if needed
-		if (answers.docpadPlugin === false && editions.length > 1) {
-			const sourceTestPathExists = await exists(sourceTestPath)
-			const sourceTestEntryExists = await exists(sourceTestEntry)
-			if (!sourceTestPathExists) {
+		if (answers.docpadPlugin === false) {
+			if ((await exists(sourceEdition.testPath)) === false) {
 				// edition entry doesn't exist, but the root entry does
-				if (sourceTestEntryExists) {
-					await rename(sourceTestEntry, sourceTestPath)
+				if (await exists(sourceEdition.test)) {
+					await rename(sourceEdition.test, sourceEdition.testPath)
 				}
 				// edition entry doesn't exist, so create a basic test file
 				else {
-					await write(sourceTestPath, [
+					await write(sourceEdition.testPath, [
 						"'use strict'",
 						'',
 						"const {equal} = require('assert-helpers')",
@@ -283,38 +249,48 @@ async function scaffoldEditions (state) {
 
 		// setup main and test paths
 		if (state.useEditionAutoloader) {
-			const sourceTestPathExists = await exists(sourceTestPath)
 			// this is the case for any language that requires compilation
 			await write('index.js', [
 				"'use strict'",
 				'',
-				`/** @type {typeof import("./${sourceMainPath}") } */`,
+				`/** @type {typeof import("./${sourceEdition.mainPath}") } */`,
 				"module.exports = require('editions').requirePackage(__dirname, require)",
 				''
 			].join('\n'))
 			packageData.main = 'index.js'
-			if (sourceTestPathExists) {
+
+			// don't both with docpad plugins
+			if (answers.docpadPlugin === false) {
 				await write('test.js', [
 					"'use strict'",
 					'',
-					`/** @type {typeof import("./${sourceTestPath}") } */`,
-					`module.exports = require('editions').requirePackage(__dirname, require, '${answers.testEntry}')`,
+					`/** @type {typeof import("./${sourceEdition.testPath}") } */`,
+					`module.exports = require('editions').requirePackage(__dirname, require, '${compiledEdition.test}')`,
 					''
 				].join('\n'))
 				state.test = 'test.js'
 			}
 		}
+		// delete the edition autoloader if it is not needed
 		else {
-			await unlink('index.js')
-			await unlink('test.js')
-			packageData.main = nodeMainPath
-			state.test = nodeTestPath
+			if (await contains('index.js', 'requirePackage')) {
+				await unlink('index.js')
+			}
+			if (await contains('test.js', 'requirePackage')) {
+				await unlink('test.js')
+			}
+
+			packageData.main = compiledEdition.mainPath
+			state.test = compiledEdition.testPath
 		}
 
 		// browser path
 		if (answers.browser) {
-			const browserEdition = editions.find((edition) => edition.engines && edition.engines.browsers) || sourceEdition
-			packageData.browser = pathUtil.join(browserEdition.directory || '.', browserEdition.entry)
+			const browserEdition = state.browserEdition || sourceEdition
+			packageData.browser = pathUtil.join(browserEdition.directory || '.', browserEdition.main)
+		}
+		else {
+			delete packageData.browser
 		}
 
 		// log
@@ -345,7 +321,7 @@ async function updateRuntime (state) {
 
 	/** @type {Object.<string, boolean | string>} */
 	const packages = {
-		'projectz': 'dev',
+		'projectz': packageData.name === 'projectz' ? false : 'dev',
 		'assert-helpers': false,
 		'joe': false,
 		'joe-reporter-console': false,
@@ -356,8 +332,19 @@ async function updateRuntime (state) {
 		'babel-core': false,
 		'babel-preset-es2015': false,
 		'babel-preset-env': false,
+		'@babel/cli': false,
+		'@babel/core': false,
+		'@babel/preset-env': false,
+		'@babel/preset-typescript': false,
+		'@babel/plugin-proposal-class-properties': false,
+		'@babel/plugin-proposal-object-rest-spread': false,
+		'typescript': false,
+		'typescript-eslint-parser': false,
 		'valid-directory': false,
 		'documentation': false,
+		'jsdoc': false,
+		'minami': false,
+		'typedoc': false,
 		'flow-bin': false,
 		'coffee-script': false,
 		'yuidocjs': false,
@@ -383,7 +370,7 @@ async function updateRuntime (state) {
 		{
 			'our:setup:npm': 'npm install',
 			'our:clean': 'rm -Rf ./docs ./edition* ./es2015 ./es5 ./out',
-			'our:meta:projectz': 'projectz compile',
+			'our:meta:projectz': packageData.name === 'projectz' ? './bin.js compile' : 'projectz compile',
 			'our:test': 'npm run our:verify && npm test',
 			'our:release:prepare': 'npm run our:clean && npm run our:compile && npm run our:test && npm run our:meta',
 			'test': `node --harmony ./${state.test} --joe-reporter=console`
@@ -402,18 +389,13 @@ async function updateRuntime (state) {
 			}
 	)
 
-	// merge in editions[scripts]
-	if (state.editions.length) {
-		Object.assign(state.scripts, ...state.editions.map((edition) => edition.scripts || {}))
-	}
-
 	// add the various scripts
 	if (answers.docpadPlugin) {
 		packages['docpad-baseplugin'] = true
 		packages['docpad-plugintester'] = packages.docpad = 'dev'
 		state.scripts.test = 'docpad-plugintester'
 		if (packageData.peerDependencies) {
-			// it is readded later
+			// it is readded later, @todo why?
 			delete packageData.peerDependencies.docpad
 		}
 	}
@@ -430,15 +412,22 @@ async function updateRuntime (state) {
 		packages.coffeelint = 'dev'
 		state.scripts['our:verify:coffeelint'] = `coffeelint ./${answers.sourceDirectory}`
 	}
-	if (answers.languages.has('esnext')) {
+	if (answers.languages.has('esnext') || answers.languages.has('typescript')) {
 		packages.eslint = 'dev'
-		state.scripts['our:verify:eslint'] = `eslint --fix ./${answers.sourceDirectory}`
+		state.scripts['our:verify:eslint'] = `eslint --fix ./${answers.sourceDirectory}/**`
+	}
+	if (answers.languages.has('typescript')) {
+		packages.typescript = packages['typescript-eslint-parser'] = 'dev'
 	}
 	if (answers.docs) {
-		if (answers.language === 'coffescript') {
+		if (answers.language === 'typescript') {
+			packages.typedoc = 'dev'
+			state.scripts['our:meta:docs'] = `typedoc --name "$npm_package_name" --readme ./README.md --out ./docs ./${answers.sourceDirectory}`
+		}
+		else if (answers.language === 'coffescript') {
 			if (packageData.devDependencies.biscotto) {
 				packages.biscotto = 'dev'
-				state.scripts['our:meta:biscotto'] = `biscotto -n "$npm_package_title" --title "$npm_package_title API Documentation" --readme README.md --output-dir ./docs ./${answers.sourceDirectory} - LICENSE.md HISTORY.md`
+				state.scripts['our:meta:biscotto'] = `biscotto -n "$npm_package_name" --title "$npm_package_name API Documentation" --readme README.md --output-dir ./docs ./${answers.sourceDirectory} - LICENSE.md HISTORY.md`
 			}
 			else {
 				packages.yuidocjs = 'dev'
@@ -446,8 +435,9 @@ async function updateRuntime (state) {
 			}
 		}
 		else if (answers.language === 'esnext') {
-			packages.documentation = 'dev'
-			state.scripts['our:meta:docs'] = `documentation build -f html -o ./docs -g --shallow ./${answers.sourceDirectory}/**.js`
+			packages.jsdoc = 'dev'
+			packages.minami = 'dev'
+			state.scripts['our:meta:docs'] = 'rm -Rf ./docs && jsdoc --recurse --pedantic --access all --destination ./docs --package ./package.json --readme ./README.md --template ./node_modules/minami ./source && mv ./docs/$npm_package_name/$npm_package_version/* ./docs/ && rm -Rf ./docs/$npm_package_name/$npm_package_version'
 		}
 	}
 	if (answers.flowtype) {
@@ -455,17 +445,25 @@ async function updateRuntime (state) {
 		state.scripts['our:verify:flow'] = 'flow check'
 	}
 	if (answers.babel) {
-		packages['babel-cli'] = packages['babel-preset-env'] = 'dev'
+		packages['@babel/core'] = packages['@babel/cli'] = packages['@babel/preset-env'] = 'dev'
+		if (answers.language === 'typescript') {
+			packages['@babel/preset-typescript'] =
+				packages['@babel/plugin-proposal-class-properties'] =
+				packages['@babel/plugin-proposal-object-rest-spread'] =
+				'dev'
+		}
 	}
-	if (answers.deploy === 'surge') {
+	if (answers.deploy) {
+		if (answers.deploy === 'surge') {
+			packages.surge = 'dev'
+			state.scripts['my:deploy'] = `surge ./${answers.deployDirectory}`
+		}
+		else if (answers.deploy.startsWith('now')) {
+			packages.now = 'dev'
+		}
+	}
+	if (answers.docs) {
 		packages.surge = 'dev'
-		state.scripts['my:deploy'] = `surge ./${answers.deployDirectory}`
-	}
-	else if (answers.docs) {
-		packages.surge = 'dev'
-	}
-	if (answers.deploy && answers.deploy.startsWith('now')) {
-		packages.now = 'dev'
 	}
 	if (!answers.docpadPlugin && !answers.website) {
 		packages.joe = packages['joe-reporter-console'] = packages['assert-helpers'] = 'dev'
@@ -533,9 +531,37 @@ async function updateRuntime (state) {
 		'.jshintrc',
 		'.jscrc',
 		'docpad-setup.sh',
-		'.babelrc'
+		'.babelrc',
+		'tsconfig.json'
 	].map((file) => unlink(file)))
 	status('...removed old files')
+
+	if (answers.language === 'typescript') {
+		status('writing tsconfig file...')
+		const tsconfig = {
+			compilerOptions: {
+				// Target latest version of ECMAScript.
+				target: 'esnext',
+				// Search under node_modules for non-relative imports.
+				moduleResolution: 'node',
+				// Process & infer types from .js files.
+				allowJs: true,
+				// Don't emit; allow Babel to transform files.
+				noEmit: true,
+				// Enable strictest settings like strictNullChecks & noImplicitAny.
+				strict: true,
+				// Disallow features that require cross-file information for emit.
+				isolatedModules: true,
+				// Import non-ES modules as default imports.
+				esModuleInterop: true
+			},
+			include: [
+				answers.sourceDirectory
+			]
+		}
+		await write('tsconfig.json', JSON.stringify(tsconfig, null, '  ') + '\n')
+		status('...wrote tsconfig file')
+	}
 
 	// running setup
 	status('running setup...')